requirements.txt
face_recognition==1.3.0
Pillow==9.0.0